streamlit/requirements.txt
streamlit==1.30.0
snowflake-connector-python[pandas]==3.5.0
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
